refactor: 전략 패턴 적용한 패키지 구조 변경
