+++ src/main/java/com/develop_ping/union/gathering/domain/GatheringManager.java
@@ -3,7 +3,6 @@ package com.develop_ping.union.gathering.domain;
 import com.develop_ping.union.gathering.domain.dto.request.GatheringListCommand;
 import com.develop_ping.union.gathering.domain.dto.response.GatheringInfo;
 import com.develop_ping.union.gathering.domain.entity.Gathering;
-import com.develop_ping.union.gathering.domain.strategy.GatheringSortStrategy;
 import org.springframework.data.domain.Slice;
 
 public interface GatheringManager {
