+++ src/main/java/com/develop_ping/union/gathering/domain/service/GatheringServiceImpl.java
@@ -7,8 +7,8 @@ import com.develop_ping.union.gathering.domain.dto.response.GatheringDetailInfo;
 import com.develop_ping.union.gathering.domain.dto.response.GatheringInfo;
 import com.develop_ping.union.gathering.domain.dto.response.GatheringListInfo;
 import com.develop_ping.union.gathering.domain.entity.Gathering;
-import com.develop_ping.union.gathering.domain.strategy.GatheringSortStrategy;
-import com.develop_ping.union.gathering.domain.strategy.GatheringSortStrategyFactory;
+import com.develop_ping.union.gathering.domain.GatheringSortStrategy;
+import com.develop_ping.union.gathering.domain.GatheringSortStrategyFactory;
 import com.develop_ping.union.gathering.exception.GatheringPermissionDeniedException;
 import com.develop_ping.union.party.domain.PartyManager;
 import com.develop_ping.union.party.domain.dto.PartyInfo;
