+++ src/main/java/com/develop_ping/union/gathering/infra/strategy/DistanceSortStrategy.java
@@ -1,5 +1,6 @@
-package com.develop_ping.union.gathering.domain.strategy;
+package com.develop_ping.union.gathering.infra.strategy;
 
+import com.develop_ping.union.gathering.domain.GatheringSortStrategy;
 import com.develop_ping.union.gathering.domain.SortType;
 import com.develop_ping.union.gathering.domain.dto.request.GatheringListCommand;
 import com.develop_ping.union.gathering.domain.entity.Gathering;
